Fixed namespace errors and typos
