+++ components/AccentAnalyzer/samples/AccentAnalyzerSample.xaml
@@ -1,18 +1,18 @@
-<!--  Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE file in the project root for more information.  -->
+﻿<!--  Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE file in the project root for more information.  -->
 <Page x:Class="AccentAnalyzerExperiment.Samples.AccentAnalyzerSample"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:behaviors="using:CommunityToolkit.WinUI.Behaviors"
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-      xmlns:extensions="using:CommunityToolkit.WinUI.Extensions"
+      xmlns:helpers="using:CommunityToolkit.WinUI.Helpers"
       xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
       xmlns:local="using:AccentAnalyzerExperiment.Samples"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
       mc:Ignorable="d">
 
     <Page.Resources>
-        <extensions:AccentAnalyzer x:Name="AccentAnalyzer"
-                                   Source="{x:Bind AccentedImage}" />
+        <helpers:AccentAnalyzer x:Name="AccentAnalyzer"
+                                Source="{x:Bind AccentedImage}" />
     </Page.Resources>
 
     <Grid>
