+++ components/AccentAnalyzer/src/AccentAnalyzer.cs
@@ -9,7 +9,7 @@ using System.Numerics;
 using System.Windows.Input;
 using Windows.UI;
 
-namespace CommunityToolkit.WinUI.Extensions;
+namespace CommunityToolkit.WinUI.Helpers;
 
 /// <summary>
 /// A resource that can be used to extract color palettes out of any UIElement.
@@ -31,137 +31,12 @@ public partial class AccentAnalyzer : DependencyObject
         }
     }
 
-    private UIElement? _source;
-
-    /// <summary>
-    /// Gets the <see cref="DependencyProperty"/> for the <see cref="PrimaryAccentColor"/> property.
-    /// </summary>
-    public static readonly DependencyProperty PrimaryAccentColorProperty =
-        DependencyProperty.Register(nameof(PrimaryAccentColor), typeof(Color), typeof(AccentAnalyzer), new PropertyMetadata(Colors.Transparent));
-    
-    /// <summary>
-    /// Gets the <see cref="DependencyProperty"/> for the <see cref="SecondaryAccentColor"/> property.
-    /// </summary>
-    public static readonly DependencyProperty SecondaryAccentColorProperty =
-        DependencyProperty.Register(nameof(SecondaryAccentColor), typeof(Color), typeof(AccentAnalyzer), new PropertyMetadata(Colors.Transparent));
-    
-    /// <summary>
-    /// Gets the <see cref="DependencyProperty"/> for the <see cref="TertiaryAccentColor"/> property.
-    /// </summary>
-    public static readonly DependencyProperty TertiaryAccentColorProperty =
-        DependencyProperty.Register(nameof(TertiaryAccentColor), typeof(Color), typeof(AccentAnalyzer), new PropertyMetadata(Colors.Transparent));
-    
-    /// <summary>
-    /// Gets the <see cref="DependencyProperty"/> for the <see cref="BaseColor"/> property.
-    /// </summary>
-    public static readonly DependencyProperty BaseColorProperty =
-        DependencyProperty.Register(nameof(BaseColor), typeof(Color), typeof(AccentAnalyzer), new PropertyMetadata(Colors.Transparent));
-    
-    /// <summary>
-    /// Gets the <see cref="DependencyProperty"/> for the <see cref="DominantColor"/> property.
-    /// </summary>
-    public static readonly DependencyProperty DominantColorProperty =
-        DependencyProperty.Register(nameof(DominantColor), typeof(Color), typeof(AccentAnalyzer), new PropertyMetadata(Colors.Transparent));
-    
-    /// <summary>
-    /// Gets the <see cref="DependencyProperty"/> for the <see cref="Colorfulness"/> property.
-    /// </summary>
-    public static readonly DependencyProperty ColorfulnessProperty =
-        DependencyProperty.Register(nameof(Colorfulness), typeof(float), typeof(AccentAnalyzer), new PropertyMetadata(0f));
-
     /// <summary>
     /// Initialize an instance of the <see cref="AccentAnalyzer"/> class.
     /// </summary>
     public AccentAnalyzer()
     {
         AccentUpdateCommand = new Command(UpdateAccent);
-    }
-
-    /// <summary>
-    /// Gets the primary accent color as extracted from the <see cref="Source"/>.
-    /// </summary>
-    /// <remarks>
-    /// The most "colorful" found in the image.
-    /// </remarks>
-    public Color PrimaryAccentColor
-    {
-        get => (Color)GetValue(PrimaryAccentColorProperty);
-        private set => SetValue(PrimaryAccentColorProperty, value);
-    }
-
-    /// <summary>
-    /// Gets the secondary accent color as extracted from the <see cref="Source"/>.
-    /// </summary>
-    /// <remarks>
-    /// The second most "colorful" color found in the image.
-    /// </remarks>
-    public Color SecondaryAccentColor
-    {
-        get => (Color)GetValue(SecondaryAccentColorProperty);
-        private set => SetValue(SecondaryAccentColorProperty, value);
-    }
-
-    /// <summary>
-    /// Gets the tertiary accent color as extracted from the <see cref="Source"/>.
-    /// </summary>
-    /// <remarks>
-    /// The third most "colorful" color found in the image.
-    /// </remarks>
-    public Color TertiaryAccentColor
-    {
-        get => (Color)GetValue(TertiaryAccentColorProperty);
-        private set => SetValue(TertiaryAccentColorProperty, value);
-    }
-    
-    /// <summary>
-    /// Gets the base color as extracted from the <see cref="Source"/>.
-    /// </summary>
-    /// <remarks>
-    /// The least "colorful" color found in the image.
-    /// </remarks>
-    public Color BaseColor
-    {
-        get => (Color)GetValue(BaseColorProperty);
-        private set => SetValue(BaseColorProperty, value);
-    }
-    
-    /// <summary>
-    /// Gets the dominent color as extracted from the <see cref="Source"/>.
-    /// </summary>
-    /// <remarks>
-    /// The most color that takes up the most of the image.
-    /// </remarks>
-    public Color DominantColor
-    {
-        get => (Color)GetValue(DominantColorProperty);
-        private set => SetValue(DominantColorProperty, value);
-    }
-    
-    /// <summary>
-    /// Gets the "colorfulness" of the <see cref="Source"/>.
-    /// </summary>
-    /// <remarks>
-    /// Colorfulness is defined by David Hasler and Sabine Susstrunk's paper on measuring colorfulness
-    /// <seealso href="https://infoscience.epfl.ch/server/api/core/bitstreams/77f5adab-e825-4995-92db-c9ff4cd8bf5a/content"/>
-    /// </remarks>
-    public float Colorfulness
-    {
-        get => (float)GetValue(ColorfulnessProperty);
-        set => SetValue(ColorfulnessProperty, value);
-    }
-
-    /// <summary>
-    /// Gets a command that executes an accent update.
-    /// </summary>
-    public ICommand AccentUpdateCommand { get; }
-
-    /// <summary>
-    /// Gets or sets the <see cref="UIElement"/>
-    /// </summary>
-    public UIElement? Source
-    {
-        get => _source;
-        set => SetSource(value);
     }
 
     /// <summary>
@@ -236,26 +111,5 @@ public partial class AccentAnalyzer : DependencyObject
 
         // Set the various properties from the UI thread
         UpdateAccentProperties(primary, secondary, tertiary, baseColor, dominantColor, overallColorfulness);
-    }
-
-    private void UpdateAccentProperties(Color primary, Color secondary, Color tertiary, Color baseColor, Color dominantColor, float colorfulness)
-    {
-        DispatcherQueue.GetForCurrentThread().TryEnqueue(() =>
-        {
-            PrimaryAccentColor = primary;
-            SecondaryAccentColor = secondary;
-            TertiaryAccentColor = tertiary;
-            DominantColor = dominantColor;
-            BaseColor = baseColor;
-            Colorfulness = colorfulness;
-        });
-    }
-
-    private void SetSource(UIElement? source)
-    {
-        _source = source;
-
-        // If true, calculate the accent color immediately.
-        _ = UpdateAccentAsync();
     }
 }
